qspi: Use Description list for steps
In Org mode, we have variety of lists to choose from.  For steps, we
can use `1)` or `A)`.  But unfortunately the current HTML exporter
doesn't convert `A)` to `<li type="A">`.

Instead, use description list with `::`, which will be converted
`<dd>`.  This also works well with the ODT exporter as well.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>